### 2021-04-16 #54 @MathieuBo
[Re] Spread of alpha-synuclein pathology through the brain connectome is modulated by selective vulnerability and predicted by network analysis
**Original article:** https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6662627/

**Scientific domain:** Neuroscience, Neurodegeneration
**Suggested editor:**  @rougier (until we find someone else?)

This is a successful replication of the [Henderson et al. 2019](https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6662627/) paper related to the spread of alpha-synuclein through the brain connectome. The code is entirely in Python3. I checked the editor/reviewer list but I am not sure who is an expert on neurodegeneration. However, the model is mostly based on a graph theory.

Mathieu Bourdenx and Thomas Paul
