### 2021-04-23 bfa161b
Ajout de Pipeline_Henrich.py & Henrich_processing.py La Pipeline permet de lancer les prédictions sur le dataset d'Henrich et al. après avoir créé une matrice de connectivité et modifié le dataset d'Henrich de sorte à avoir une compatibilité avec la matrice de connectivité. Henrich_processing renvoie W, Path_data, ROI_names et conn_names. grp_mean non nécessaire car n'ayant qu'un groupe de souris/timepoint.
Les autres analyses (Contrôles, Robustesse, Modèle Itératif) n'ont pas été pour le moment utilisées et pourraient nécessiter quelques modifications préalables.
