+++ Pipeline_Henrich.py
@@ -8,6 +8,7 @@ from statsmodels.stats.multitest import multipletests
 
 np.seterr(divide='ignore')  # Hide Runtime warning regarding log(0) = -inf
 
+import Henrich_processing
 import process_files
 import Laplacian
 import fitfunctions
@@ -19,6 +20,7 @@ plt.rc('xtick', labelsize=14)
 plt.rc('ytick', labelsize=14)
 
 # Directories creation in the working directory
+os.chdir('C:\\Users\\thoma\\Documents\\M1_Neurasmus\\NeuroBIM_M1\\Internship\\Henrich_et_al_Data\\Final_Files')
 os.makedirs("../output", exist_ok=True)
 os.makedirs("../plots", exist_ok=True)
 os.makedirs("../Iterative_Model/plots/Density_vs_residuals", exist_ok=True)
@@ -29,28 +31,33 @@ for directory in directories:
 
 class DataManager(object):
     # Initialization
-    def __init__(self, exp_data, synuclein, timepoints, seed, use_expression_values=None):
+    def __init__(self, exp_data, synuclein, timepoints, seed, condition, use_expression_values=None):
 
         self.seed = seed
-
+        self.condition = condition
         # Connectivity tables
-        self.connectivity_ipsi = pd.read_csv("./Data83018/connectivity_ipsi.csv", index_col=0)
+        self.connectivity_ipsi = pd.read_csv("./Connectivity_Map_Henrich_ipsi.csv",
+                                            header=0, index_col=0)
 
-        self.connectivity_contra = pd.read_csv("./Data83018/connectivity_contra.csv", index_col=0)
+        self.connectivity_contra = pd.read_csv("./Connectivity_Map_Henrich_contra.csv",
+                                              header=0, index_col=0)
 
-        self.W, self.path_data, self.conn_names, self.ROInames = process_files.process_pathdata(exp_data=exp_data,
+        self.W, self.path_data, self.conn_names, self.ROInames = Henrich_processing.process_pathdata(exp_data=exp_data,
                                                                                                 connectivity_contra=self.connectivity_contra,
-                                                                                                connectivity_ipsi=self.connectivity_ipsi)
+                                                                                                connectivity_ipsi=self.connectivity_ipsi,
+                                                                                                condition=self.condition)
 
         self.synuclein = process_files.process_gene_expression_data(expression=synuclein, roi_names=self.ROInames)
 
         self.timepoints = timepoints
 
+        self.condition = condition
+
         self.l_out = None
 
-        self.grp_mean = process_files.mean_pathology(timepoints=timepoints, path_data=self.path_data)
+        #self.grp_mean removed in that code
 
-        self.c_rng = np.linspace(start=0.01, stop=100, num=1000)
+        self.c_rng = np.linspace(start=0.01, stop=10, num=100)
 
         self.use_expression_values = use_expression_values
 
@@ -72,7 +79,7 @@ class DataManager(object):
 
     def find_best_c_and_r(self):
 
-        c_Grp, r = fitfunctions.c_fit(log_path=np.log10(self.grp_mean),
+        c_Grp, r = fitfunctions.c_fit(log_path=np.log10(self.path_data),
                                       L_out=self.l_out,
                                       tp=timepoints,
                                       seed=self.seed,
@@ -86,9 +93,9 @@ class DataManager(object):
[…]
