+++ Pipeline_Henrich.py
@@ -1,405 +1,0 @@
-import pandas as pd
-import numpy as np
-import seaborn as sns
-import matplotlib.pyplot as plt
-import os
-from scipy.stats import pearsonr, linregress
-from statsmodels.stats.multitest import multipletests
-
-np.seterr(divide='ignore')  # Hide Runtime warning regarding log(0) = -inf
-
-import Henrich_processing
-import process_files
-import Laplacian
-import fitfunctions
-import Robustness_Stability
-import summative_model
-
-# Plot settings
-plt.rc('xtick', labelsize=14)
-plt.rc('ytick', labelsize=14)
-
-# Directories creation in the working directory
-os.chdir('C:\\Users\\thoma\\Documents\\M1_Neurasmus\\NeuroBIM_M1\\Internship\\Henrich_et_al_Data\\Final_Files')
-os.makedirs("../output", exist_ok=True)
-os.makedirs("../plots", exist_ok=True)
-os.makedirs("../Iterative_Model/plots/Density_vs_residuals", exist_ok=True)
-directories = ["Density_vs_residuals", "Heatmap_Predictions", "Fits(c)",
-               "Model_Robustness", "Stab_Grp", "Stab_Ind", "Stab_Ind_Sliding"]
-for directory in directories:
-    os.makedirs("../plots/{}".format(directory), exist_ok=True)
-
-class DataManager(object):
-    # Initialization
-    def __init__(self, exp_data, synuclein, timepoints, seed, condition, use_expression_values=None):
-
-        self.seed = seed
-        self.condition = condition
-        # Connectivity tables
-        self.connectivity_ipsi = pd.read_csv("./Connectivity_Map_Henrich_ipsi.csv",
-                                            header=0, index_col=0)
-
-        self.connectivity_contra = pd.read_csv("./Connectivity_Map_Henrich_contra.csv",
-                                              header=0, index_col=0)
-
-        self.W, self.path_data, self.conn_names, self.ROInames = Henrich_processing.process_pathdata(exp_data=exp_data,
-                                                                                                connectivity_contra=self.connectivity_contra,
-                                                                                                connectivity_ipsi=self.connectivity_ipsi,
-                                                                                                condition=self.condition)
-
-        self.synuclein = process_files.process_gene_expression_data(expression=synuclein, roi_names=self.ROInames)
-
-        self.timepoints = timepoints
-
-        self.condition = condition
-
-        self.l_out = None
-
-        #self.grp_mean removed in that code
-
-        self.c_rng = np.linspace(start=0.01, stop=10, num=100)
-
-        self.use_expression_values = use_expression_values
-
-        print('Data Manager initialized\n')
-
-    # Main Functions
-    def compute_graph(self):
-
-        if self.use_expression_values:
-
-            self.l_out = Laplacian.get_laplacian(adj_mat=self.W, expression_data=self.synuclein)
-            print('Using a syn expression values')
-
-        else:
-
-            self.l_out = Laplacian.get_laplacian(adj_mat=self.W)
-
-        print('Graph computed - Laplacian matrix created\n')
-
-    def find_best_c_and_r(self):
-
-        c_Grp, r = fitfunctions.c_fit(log_path=np.log10(self.path_data),
-                                      L_out=self.l_out,
-                                      tp=timepoints,
-                                      seed=self.seed,
-                                      c_rng=self.c_rng,
-                                      roi_names=self.ROInames)
-
-        return c_Grp, r
-
[…]
