### 2021-05-07 #54 @MathieuBo
@rougier, I have been looking for names of people adopting similar approaches:
- Henry Kennedy (academic editor of a related paper): henry.kennedy /at inserm.fr
- Alain Dagher (author of paper on related subject): alain.dagher /at mcgill.ca
- Bratislav Misic (author of paper on related subject): bratislav.misic /at mcgill.ca
- Mike Henderson (author of the paper that we replicated - conflict?):  Michael.Henderson /at vai.org
- Ashish Raj (author of the first model): ashish.raj /at ucsf.edu
- Jacob Vogel (author of paper using related models): jacob.vogel /at mail.mcgill.ca
