### 2021-05-17 #54 @rougier
Dear @ReScience/editors we need your help to edit this submission. Author has already proposed several potential reviewers. I think the domain is far from any of us and I cannot edit it because of conflict of interest. Can anyone edit it (link: https://github.com/ReScience/submissions/issues/54)
