### 2021-05-31 #54 @khinsen
@bmisic @Raj-Lab-UCSF You have been suggested as reviewers for a submission to the journal "ReScience C". Would be interested in doing this review? Please read on for the details.

ReScience C is a journal dedicated to replications in computational science. The submission I would like to invite you to review is a replication of a 2020 publication by Mike Henderson et al. entitled "α-Synuclein pathology spread through the brain connectome is modulated by selective vulnerability and predicted by network analysis". The goal of a replication article is to (1) explore if and to what degree published results can be replicated using different computational tools and (2) provide a reproducible implementation of the applied methods using Open Source software, for easier future validation and reuse. The reviewers' job is to verify those two aspects of the submission. This is much more technical than reviewing for traditional scientific journals.

Note also that ReScience C practices an extreme form of open reviewing: the whole submission and reviewing process happens in public, such that anyone can see it and even jump in. In fact, this message I am addressing to you is part of this open reviewing process.
