### 2021-05-31 #54 @khinsen
@MathieuBo In the meantime, a question based on my own reading of your article. You state that the original work already comes with code and data. It isn't clear from your description if you used this code or not, and if you did, to what degree: as a source of inspiration for your own code, or only as a black-box result generator for testing your own implementation. Could you explain?
