### 2021-05-31 #54 @MathieuBo
Thanks for looking into our submission @khinsen!
To answer your question, it was a bit of both. We mostly used it as an inspiration for our code (for example, we kept some function names similar to ease comparison) but the precise implementation diverges in some ways (mostly related to Python vs R). We also used it to generate results and compare them with the results of our code.
Is that clearer?
