### 2021-05-31 #54 @khinsen
Thanks @MathieuBo! You should add a paragraph on this to your, so that readers know what exactly was the subject of the replication. There's nothing wrong in taking inspiration from existing code, but it means that if there's a mistake in that code, it could propagate into yours, and readers should be aware of that possibility.
