### 2021-06-18 #54 @illdopejake
Hi all -- checking in as a Reviewer. I'll need to carefully re-read the original paper. As per instructions, I'll give some feedback about the replication manuscript with respect to whether claims of replication have been substantiated. After, I'll run the code provided to ensure I can reproduce everything, and I'll check in during that process if any issues arise. Will try to have this done in 2 weeks, if that's alright.

If I missed anything or not understanding, please do correct me / guide me.
