### 2021-07-05 #54 @illdopejake
Dear all,

Apologies for the delay. I've read through the original Henderson et al. manuscript and the Paul & Bourdenx ReScienceC manuscript. I also took a try of reproducing the code locally. I have some feedback as it pertains to this initial experience with the manuscript and code.

**Regarding the manuscript**:
I found the manuscript to be well written, conveying in appropriate detail all information necessary to understand the replication effort without the need for re-reading the original manuscript. The methods and results were also well contextualized, and all key figures, tables and points necessary to convey the details of the replication were present. However, I did want to bring up these two minor points:

* The mouse brain plots in Figure 2D appear not to be identical to those in the original Henderson et al. papers. This is hard to understand, since the statistical tests on that same data perfectly replicate the results described by Henderson et al. Whether the difference is actually present or simply illusory, I could imagine a few possibilities that may contribute to this discrepancy:
	a) The color scale has a different range. It is -1 to 1 in Henderson et al., and -2 to 1.5 in the present work.
	b) The rendering software differs, and the slices appear not to be perfectly harmonized with the figure in Henderson et al.

* Related to a comment below, I am curious whether there were measures that were necessary during the present replication that were needed to ensure the replication was perfect, that were not mentioned in the original Henderson et al. manuscript. In other words, did the authors need to "figure out" any extra steps during data preprocessing or statistical analysis in order to obtain results identical to Henderson et al. If so, it would be of interest to describe in the manuscript any extra steps that were taken.

==========

**First foray into the code**
I perused the github repo, cloned it, set up a virtual Python 3 env with only the necessary libraries, and ran through Pipeline.py. I uncommented functions in lines 365 till the end in order to see if all functions worked. Most of the code worked beautifully, but I had some comments and I did eventually hit some errors/discrepancy during my attempt at reproduction. I will list comments, suggestions and questions here, followed by reproduction errors:

**Comments and suggestions**

* The github repo has a number of jupyter notebooks that are not mentioned in the documentation. One of these involves generation of Figures for the present manuscript, whereas others involve data analysis both reported and not reported. Were these notebooks used in the process of replication, and is there anything within them that warrants mention? Are they simply supplemental analyses to demonstrate use cases of the code? Will they be removed or retained in the future? It would be useful if a guide to or explanation of these notebooks was included or at least mentioned in the documentation.

* At first, I considered playing around with the code in Pipeline.py interactively, so I started loading some of the information into a jupyter notebook. However, I was unable to proceed as it appears that `timepoints` is used as a global variable within the function `dm.find_best_c_and_r()` (invoked in line 358 of Pipeline.py). I'm not sure if this is intentional and there may be other instances. The authors may not be concerned with this as it does not impede replication or execution of Pipeline.py as a script, but I figured it might be worth mentioning.

* At some point, I was guided by an error that I needed to change `WindowsError` to `OSError` in line 83 of summative_model.py (perhaps this is necessary elsewhere?). This may warrant mention in the documentation, though the error message was helpful.

* The output folders for Pipeline.py appear to be hardcoded into a directory one above (`../`). This is not where the output livesin the git repo. It is not necessary to give an argument for directory, but I should note that when I tried changing the locations of output directories (lines 22-28 in Pipeline.py), this caused an error later in the script.

**Reproduction failures**

* When performing the task “Plotting the iCPu Fit versus Fits of random seed regions…” , the result was “iCPu seed is the 56.0th percentile”. Perhaps I'm misunderstanding what this means, but it seems to be at odds with the results in the paper, and does not seem consistent with the plots that were generated (which look like the plots in manuscript).

* Similar experience  while “Plotting the adjacency matrix Fit versus Fits of random adjacency matrices…”, the result was “iCPu seed is the 58.0th percentile”.

* Plotting the non-shuffled pathology Fit versus shuffled pathology fits, I ran into the following error:
....Paul_ReScienceC_paper/revC/lib/python3.7/site-packages/scipy/stats/stats.py:4040: RuntimeWarning: invalid value encountered in subtract                                                 | 0/3 [00:00<?, ?it/s]
  ym = y.astype(dtype) - ymean
  0%|                                                                                                                                                                                                                         | 0/3 [00:00<?, ?it/s]
  5%|██████████▊                                                                                                                                                                                                     | 3/58 [00:00<00:00, 68.01it/s]
Traceback (most recent call last):
  File "Pipeline.py", line 389, in <module>
    dm.compute_stability(Sliding_Window=None)
  File "Pipeline.py", line 291, in compute_stability
    suffix=suffix, seed=seed)
  File "/Users/jacobv/Documents/Papers/Reviews/Paul_ReScienceC_paper/replication/Robustness_Stability.py", line 235, in stability
    roi_names=ROInames)
  File "/Users/jacobv/Documents/Papers/Reviews/Paul_ReScienceC_paper/replication/fitfunctions.py", line 78, in c_fit
    r, _ = stats.pearsonr(exp_val, predict_val)
  File "/Users/jacobv/Documents/Papers/Reviews/Paul_ReScienceC_paper/revC/lib/python3.7/site-packages/scipy/stats/stats.py", line 4046, in pearsonr
    normym = linalg.norm(ym)
  File "/Users/jacobv/Documents/Papers/Reviews/Paul_ReScienceC_paper/revC/lib/python3.7/site-packages/scipy/linalg/misc.py", line 145, in norm
    a = np.asarray_chkfinite(a)
  File "/Users/jacobv/Documents/Papers/Reviews/Paul_ReScienceC_paper/revC/lib/python3.7/site-packages/numpy/lib/function_base.py", line 489, in asarray_chkfinite
    "array must not contain infs or NaNs")
ValueError: array must not contain infs or NaNs

**Conclusion**
I would be happy to work with the authors to figure out the source of this error. I would ideally like to change some of the input arguments to check the robustness of the pipeline, but will wait till the above error is resolved. Please let me know if there are other aspects of the reproduction that the authors or editors feel require review.

--Jake
