### 2021-07-07 #54 @MathieuBo
Thanks @illdopejake for taking the time and the constructive comments!
I will work on these issues as soon as possible.
