+++ Pipeline.py
@@ -12,26 +12,19 @@ import process_files
 import Laplacian
 import fitfunctions
 import Robustness_Stability
-import summative_model
+
 
 # Plot settings
-plt.rc('xtick', labelsize=14)
-plt.rc('ytick', labelsize=14)
+plt.rc('xtick', labelsize=18)
+plt.rc('ytick', labelsize=18)
 
-# Directories creation in the working directory
-os.makedirs("../output", exist_ok=True)
-os.makedirs("../plots", exist_ok=True)
-os.makedirs("../Iterative_Model/plots/Density_vs_residuals", exist_ok=True)
-directories = ["Density_vs_residuals", "Heatmap_Predictions", "Fits(c)",
-               "Model_Robustness", "Stab_Grp", "Stab_Ind", "Stab_Ind_Sliding"]
-for directory in directories:
-    os.makedirs("../plots/{}".format(directory), exist_ok=True)
 
 class DataManager(object):
     # Initialization
-    def __init__(self, exp_data, synuclein, timepoints, seed, use_expression_values=None):
+    def __init__(self, exp_data, synuclein, timepoints, seed, output_path, use_expression_values=None, file_format='png'):
 
         self.seed = seed
+        self.output_path = output_path
 
         # Connectivity tables
         self.connectivity_ipsi = pd.read_csv("./Data83018/connectivity_ipsi.csv", index_col=0)
@@ -54,7 +47,29 @@ class DataManager(object):
 
         self.use_expression_values = use_expression_values
 
+        self.format = file_format
+
         print('Data Manager initialized\n')
+
+    # Initialization of the data names & directories:
+    def initialization(self):
+        """ Creates a suffix name that will be used to save the folders and the document with appropriate names"""
+        if self.use_expression_values:
+            gene_exp = '_SNCA'
+            suffix = "_{}{}".format(self.seed, gene_exp)
+        else:
+            suffix = "_{}".format(self.seed)
+
+        print("Name of the folder containing the output graphs is:  ", suffix)
+
+        os.makedirs("{}/{}/Tables".format(self.output_path, suffix), exist_ok=True)
+        os.makedirs("{}/{}/plots".format(self.output_path, suffix), exist_ok=True)
+        directories = ["Density_vs_residuals", "Heatmap_Predictions", "Fits(c)",
+                       "Model_Robustness"]
+        for directory in directories:
+            os.makedirs("{}/{}/plots/{}".format(self.output_path, suffix, directory), exist_ok=True)
+
+        return suffix
 
     # Main Functions
     def compute_graph(self):
@@ -74,7 +89,7 @@ class DataManager(object):
 
         c_Grp, r = fitfunctions.c_fit(log_path=np.log10(self.grp_mean),
                                       L_out=self.l_out,
-                                      tp=timepoints,
+                                      tp=self.timepoints,
                                       seed=self.seed,
                                       c_rng=self.c_rng,
                                       roi_names=self.ROInames)
@@ -83,42 +98,23 @@ class DataManager(object):
 
     def predict_pathology(self, c_Grp):
 
-        # Initialization
-        if self.use_expression_values:
-            gene_exp = '_SNCA'
-            suffix = "_{}{}".format(self.seed, gene_exp)
-        else:
-            suffix = "_{}".format(self.seed)
-
         Xo = fitfunctions.make_Xo(ROI=self.seed, ROInames=self.ROInames)
-
-        print("suffix is ", suffix)
 
         Xt_Grp = [fitfunctions.predict_Lout(self.l_out, Xo, c_Grp, i) for i in timepoints]
 
[…]
