+++ summative_model.py
@@ -1,185 +1,0 @@
-import os
-import numpy as np
-import pandas as pd
-import matplotlib.pyplot as plt
-from scipy.linalg import expm
-from scipy.stats import pearsonr, linregress
-import seaborn as sns
-from statsmodels.stats.multitest import multipletests
-import fitfunctions
-
-
-def extract_c_and_r_iter(log_path, L_out, tp, seed, c_rng, roi_names):
-    """
-    Extract each c and r values corresponding while iterating on c.
-    ---
-    Inputs:
-    log_path --> log10 of grp_mean. Grp_mean is the Dataframe with mean pathology per group, timepoints and regions
-    L_out --> Laplacian matrice, array
-    tp --> Timepoint, list
-    c_rng --> Constant to tune the time scale
-    roi_names --> ROInames
-    ---
-    Outputs:
-    extracted --> Panda DataFrame with Multi-indexed columns. Contains for each timepoint c values and corresponding
-    r values
-    """
-    global best_c_per_mpi1, best_c_per_mpi3
-    Xo = fitfunctions.make_Xo(seed, roi_names)
-    # Exclusion mask; we do not count the regions with 0 path
-    mask = log_path != -np.inf
-    # Compute fit at each time point for range of time
-    multi = []
-    for time in tp:
-        for output in ["c", "r"]:
-            multi.append((str(time), output))
-
-    col = pd.MultiIndex.from_tuples(multi, names=["MPI", "Condition"])
-    c_idx = [i for i in range(0,len(c_rng))]
-    extracted = pd.DataFrame(0, index=c_idx, columns=col)
-    for time in range(0, len(tp)):
-
-        for c_idx, c_value in enumerate(c_rng):
-            exp_val = log_path.iloc[:, time][mask.iloc[:, time]].values
-            if tp[time] == 1:
-                Xt_1 = np.dot(expm(-L_out * c_value * 1), Xo) + 0
-                predict_val = np.log10(Xt_1[mask.iloc[:, time]])
-                r, _ = pearsonr(exp_val, predict_val)
-                extracted[str(tp[time]), "c"].loc[c_idx] = + c_value
-                extracted[str(tp[time]), "r"].loc[c_idx] = + r
-                idx = np.where(extracted[str(tp[time]), "r"] == np.max(extracted[str(tp[time]), "r"]))[0][0]
-                best_c_per_mpi1 = extracted[str(tp[time]), "c"][idx]
-
-            if tp[time] == 3:
-                Xt_3 = np.dot(expm(-L_out * c_value * 3), np.dot(expm(-L_out * best_c_per_mpi1 * 1), Xo) + 0) + np.dot(expm(-L_out * best_c_per_mpi1 * 1), Xo) + 0
-                predict_val = np.log10(Xt_3[mask.iloc[:, time]])
-                r, _ = pearsonr(exp_val, predict_val)
-                extracted[str(tp[time]), "c"].loc[c_idx] = + c_value
-                extracted[str(tp[time]), "r"].loc[c_idx] = + r
-                idx = np.where(extracted[str(tp[time]), "r"] == np.max(extracted[str(tp[time]), "r"]))[0][0]
-                best_c_per_mpi3 = extracted[str(tp[time]), "c"][idx]
-
-            if tp[time] == 6:
-                Xt_6 = np.dot(expm(-L_out * c_value * 6), (np.dot(expm(-L_out * best_c_per_mpi3 * 3), np.dot(expm(-L_out * best_c_per_mpi1 * 1), Xo) + 0) + np.dot(expm(-L_out * best_c_per_mpi1 * 1), Xo) + 0)) \
[…]
