### 2021-07-30 #54 @MathieuBo
**Authors Response (AR)**: Dear @illdopejake,
Thank you for reviewing our reproducibility report! Please find below answers to your comments.

>Dear all,
Apologies for the delay. I've read through the original Henderson et al. manuscript and the Paul & Bourdenx ReScienceC manuscript. I also took a try of reproducing the code locally. I have some feedback as it pertains to this initial experience with the manuscript and code.
Regarding the manuscript:
I found the manuscript to be well written, conveying in appropriate detail all information necessary to understand the replication effort without the need for re-reading the original manuscript. The methods and results were also well contextualized, and all key figures, tables and points necessary to convey the details of the replication were present. However, I did want to bring up these two minor points:
•	The mouse brain plots in Figure 2D appear not to be identical to those in the original Henderson et al. papers. This is hard to understand, since the statistical tests on that same data perfectly replicate the results described by Henderson et al. Whether the difference is actually present or simply illusory, I could imagine a few possibilities that may contribute to this discrepancy:
>1)	a) The color scale has a different range. It is -1 to 1 in Henderson et al., and -2 to 1.5 in the present work.
>2)	b) The rendering software differs, and the slices appear not to be perfectly harmonized with the figure in Henderson et al.

**AR**: Thanks a lot for raising that issue. In Figure 2D, we plotted the residuals obtained for every brain region after averaging the residuals from the three different timepoint (1, 3 and 6 months post-injection – MPI). As you mentioned, our code perfectly replicated the original publication. For confirmation, we here plotted the residuals obtained using our code and using the original code.

![Residual plot](https://github.com/MathieuBo/Re-Henderson-2019/blob/master/answer_to_reviewer/ResidualsHender_vs_Our_values_MPI1.png)

The difference in rendering can be explained by various reasons.

* In this replication, we used the recently developed BrainRender package that is based on the Allen Brain Mouse atlas with the Common Coordinate Framework v3. The pathology quantification performed by Henderson et al. in the original publication was based on manual annotation of 172 brain regions on 5 brain levels. The differences between the two set of annotations required several manual adjustments to match levels of granularity. In some instances, we chose not to merge several regions in order to avoid providing an exaggerated feeling of homogeneity. From the Materials & Methods section, it is unclear whether Henderson and colleagues used a rendering software or not.
* Second, BrainRender does not have a clear correspondence with the Bregma-Lambda coordinate system. We selected slices that approach as much as possible the original section but again, although the overall level was similar, there are some differences because of region delimitations.
* You are right to point out that our scale is different. From the original publication, it is unclear whether the residuals were normalized using a min/max score. When we tried to normalize between -1 and 1 using a min/max, the output was exactly similar. Hence, we preferred to keep the full range of residuals.
* Overall, despite these minute differences related to visual representation, we believe that the figure convey a similar message. The reader can easily identify regions of vulnerability: primary motor cortex, piriform cortex, amygdala region, etc.

In the revised version of the manuscript, we have now added a sentence mentioning the differences due to the set of annotations. Also, if you think it would be useful, we could include the plot shown above in the replication.

>	Related to a comment below, I am curious whether there were measures that were necessary during the present replication that were needed to ensure the replication was perfect, that were not mentioned in the original Henderson et al. manuscript. In other words, did the authors need to "figure out" any extra steps during data preprocessing or statistical analysis in order to obtain results identical to Henderson et al. If so, it would be of interest to describe in the manuscript any extra steps that were taken.

**AR**: Except the rendering of brain slices for which we decided to use BrainRender (in order to automatize the rendering of such panels in the future), the replication required no extra steps. The available code was clear and could be followed easily.

> First foray into the code
I perused the github repo, cloned it, set up a virtual Python 3 env with only the necessary libraries, and ran through Pipeline.py. I uncommented functions in lines 365 till the end in order to see if all functions worked. Most of the code worked beautifully, but I had some comments and I did eventually hit some errors/discrepancy during my attempt at reproduction. I will list comments, suggestions and questions here, followed by reproduction errors:
Comments and suggestions

>* The github repo has a number of jupyter notebooks that are not mentioned in the documentation. One of these involves generation of Figures for the present manuscript, whereas others involve data analysis both reported and not reported. Were these notebooks used in the process of replication, and is there anything within them that warrants mention? Are they simply supplemental analyses to demonstrate use cases of the code? Will they be removed or retained in the future? It would be useful if a guide to or explanation of these notebooks was included or at least mentioned in the documentation.

**AR**: As this replication is part of a bigger study, we mistakenly left some unnecessary Jupyter Notebooks not directly related to the replication. In order to avoid concerns, we chose to clean the main branch of the repository to leave only the required files. We have also edited the documentation.

>* 	At first, I considered playing around with the code in Pipeline.py interactively, so I started loading some of the information into a jupyter notebook. However, I was unable to proceed as it appears that timepoints is used as a global variable within the function dm.find_best_c_and_r() (invoked in line 358 of Pipeline.py). I'm not sure if this is intentional and there may be other instances. The authors may not be concerned with this as it does not impede replication or execution of Pipeline.py as a script, but I figured it might be worth mentioning.

**AR**: Thank you for warning us about this. We have now corrected the code.

>*	At some point, I was guided by an error that I needed to change WindowsError to OSError in line 83 of summative_model.py (perhaps this is necessary elsewhere?). This may warrant mention in the documentation, though the error message was helpful.

**AR**: We apologize for this issue related to the fact that the repository contained unnecessary files. We have now remove this file.

>*	The output folders for Pipeline.py appear to be hardcoded into a directory one above (../). This is not where the output livesin the git repo. It is not necessary to give an argument for directory, but I should note that when I tried changing the locations of output directories (lines 22-28 in Pipeline.py), this caused an error later in the script.

**AR**: We fixed the output folders in the updated version of the code. The output path can now be chosen in the data manager. We consequently corrected the ReadMe section.

>Reproduction failures

>*	When performing the task “Plotting the iCPu Fit versus Fits of random seed regions…” , the result was “iCPu seed is the 56.0th percentile”. Perhaps I'm misunderstanding what this means, but it seems to be at odds with the results in the paper, and does not seem consistent with the plots that were generated (which look like the plots in manuscript).

**AR**: Very good point. We fixed it as returning the percentile of the true seeded region is only relevant when assessing the specificity of the model.

>*	Similar experience while “Plotting the adjacency matrix Fit versus Fits of random adjacency matrices…”, the result was “iCPu seed is the 58.0th percentile”.

**AR**: Fixed in the updated version. Here again, we removed that print as it was not relevant for this function (it came from a non-corrected copy/paste).

>*	Plotting the non-shuffled pathology Fit versus shuffled pathology fits, I ran into the following error:
....Paul_ReScienceC_paper/revC/lib/python3.7/site-packages/scipy/stats/stats.py:4040: RuntimeWarning: invalid value encountered in subtract | 0/3 [00:00<?, ?it/s]
ym = y.astype(dtype) - ymean
0%| | 0/3 [00:00<?, ?it/s]
5%|██████████▊ | 3/58 [00:00<00:00, 68.01it/s]
Traceback (most recent call last):
File "Pipeline.py", line 389, in
dm.compute_stability(Sliding_Window=None)
File "Pipeline.py", line 291, in compute_stability
suffix=suffix, seed=seed)
File "/Users/jacobv/Documents/Papers/Reviews/Paul_ReScienceC_paper/replication/Robustness_Stability.py", line 235, in stability
roi_names=ROInames)
File "/Users/jacobv/Documents/Papers/Reviews/Paul_ReScienceC_paper/replication/fitfunctions.py", line 78, in c_fit
r, _ = stats.pearsonr(exp_val, predict_val)
File "/Users/jacobv/Documents/Papers/Reviews/Paul_ReScienceC_paper/revC/lib/python3.7/site-packages/scipy/stats/stats.py", line 4046, in pearsonr
normym = linalg.norm(ym)
File "/Users/jacobv/Documents/Papers/Reviews/Paul_ReScienceC_paper/revC/lib/python3.7/site-packages/scipy/linalg/misc.py", line 145, in norm
a = np.asarray_chkfinite(a)
File "/Users/jacobv/Documents/Papers/Reviews/Paul_ReScienceC_paper/revC/lib/python3.7/site-packages/numpy/lib/function_base.py", line 489, in asarray_chkfinite
"array must not contain infs or NaNs")
ValueError: array must not contain infs or NaNs

**AR**: Here again, we sincerely apologize as this function came from a test that was not completely mature. We have now removed this function from the repository.
We could still explain why you obtained this error. The point of this function was to determine what is the minimal number of brain regions that should be studied to have stable predictions using the NDM model. To do so, it select random slices of the connectivity matrix. Unfortunately, this sometimes led to selection of regions with no pathology and later in the code to a “ValueError: array must not contain infs or NaNs” error. This could be solved by using a while True-try loop. Still, we decided to remove this part of the code as there is no direct application in the replication.

>Conclusion
>
>I would be happy to work with the authors to figure out the source of this error. I would ideally like to change some of the input arguments to check the robustness of the pipeline, but will wait till the above error is resolved. Please let me know if there are other aspects of the reproduction that the authors or editors feel require review.
--Jake

**AR**: Thanks again for your comments and pointing out the various issues. Please let us know if you would like us to modify other aspects of the code.

Best,

Paul & Mathieu
