### 2021-08-12 #54 @ejcorn
Hi all, sorry for my delay on this. I have a few small comments but the manuscript and code look great. Certainly an improvement upon our code repository and happy to see a successful replication!

**Manuscript**

- Graph theory and brain connectivity: from the perspective of matrix multiplication in linear systems, outgoing connections are the columns and incoming are the rows. This is counterintuitive but important.
- I agree with @illdopejake's reasons for why the heatmap glass mouse brains don't look identical. Unfortunately the code to generate these heatmaps is not readily available, and i don't believe it's written in python, R or matlab. We will keep BrainRender in mind for future studies.
- the control where pathology is shuffled is a nice addition

**Code**
After the authors incorporated @illdopejake's comments, I was able to run `pipeline.py` with no errors using python 3.7.6 installed with anaconda. I didn't create a new environment. This code runs much better than ours so I'm happy to see it out there! A couple of small recommendations:

- at one point the script paused and I needed to shut the figure windows in order for it to advance. You could consider printing a message that says "Close figure windows to advance.", but this is not necessary.
