### 2021-08-24 #54 @MathieuBo
Sorry for the delay to answer (August in France ...). Thanks a lot @ejcorn for reviewing our manuscript and code!
I have now edited the description of the connectivity matrix to correct for the inversion between incoming and outgoing (thanks for catching that!). I already pushed the new version of the manuscript on the repository.

Regarding the code execution, I am not able to reproduce the pause that you mention. Could you specify when that happened?

@illdopejake, thanks again for your comments. Glad the correction solved the issues.
