### 2021-08-29 #54 @ejcorn
Probably just something to do with my package versions or GUI set up. I didn't run it in a clean environment. No need to chase after this. If you're curious, here's the output:

```
Adjacency matrix: successful concatenation
Data Manager initialized

Name of the folder containing the output graphs is:   _iCPu
Graph computed - Laplacian matrix created

---------------------------------------------------
--------------NETWORK DIFFUSION MODEL--------------
---------------------------------------------------

---------------------------------------------------
Month Post Injection 1
Number of Regions used:  97
Pearson correlation coefficient 0.5588719364388353
Pvalue (non corrected) 2.695125677527694e-09
---------------------------------------------------

---------------------------------------------------
Month Post Injection 3
Number of Regions used:  113
Pearson correlation coefficient 0.6960522094835953
Pvalue (non corrected) 1.1504825644074288e-17
---------------------------------------------------

---------------------------------------------------
Month Post Injection 6
Number of Regions used:  113
Pearson correlation coefficient 0.6478470914996128
Pvalue (non corrected) 8.793043099888211e-15
---------------------------------------------------

iCPu was dropped to create the heatmap.
```

At that point, 10 figures pop up that I need to close before it will proceed to:

```
Loading of Random Seeding test:
  8%|███▎                                       | 9/116 [00:06<01:15,  1.42it/s]
```
