### 2021-08-31 3031496
Edit display plot default and adjust README
