### 2021-08-31 #54 @MathieuBo
@ejcorn , I found the issue.
Depending on how you execute the code (within an IDE or through terminal), the script was pausing after plotting awaiting for the user to close the figures before proceeding.
In order to prevent that issue, I have created a new argument ```display_plots``` in the ```DataManager``` class. By default it is set to ```False``` and the plots are saved to the output folder.
I hope that is an acceptable solution. I here prefered to not use additional librairies.

Thanks!
