### 2021-09-01 #54 @khinsen
Thanks @MathieuBo and @ejcorn for looking into what seems to me like a matplotlib backend configuration issue.

Question for @ejcorn: are you happy with the article and code as it is now? In other words, can we accept it for publication?
