### 2021-09-01 #54 @ejcorn
Yes, absolutely! I’m more than happy with the article and code. I recommend for publication.

From: Konrad Hinsen ***@***.***>
Reply-To: ReScience/submissions ***@***.***>
Date: Wednesday, September 1, 2021 at 5:35 AM
To: ReScience/submissions ***@***.***>
Cc: "Cornblath, Eli" ***@***.***>, Mention ***@***.***>
Subject: [External] Re: [ReScience/submissions] [Re] Spread of alpha-synuclein pathology through the brain connectome is modulated by selective vulnerability and predicted by network analysis (#54)

Thanks @MathieuBo<https://github.com/MathieuBo> and @ejcorn<https://github.com/ejcorn> for looking into what seems to me like a matplotlib backend configuration issue.

Question for @ejcorn<https://github.com/ejcorn>: are you happy with the article and code as it is now? In other words, can we accept it for publication?
