### 2021-09-02 #54 @khinsen
@MathieuBo I need your help for the publication process because I get LaTeX errors when trying to compile your article. Could you please accept https://github.com/MathieuBo/Re-Henderson-2019/pull/1 and compile the final version of the article? Feel free to check my changes to the metadata of course (except for the DOI, which isn't active yet so you cannot check it).
