### 2021-09-02 #54 @MathieuBo
@khinsen, I have just merged your PR, recompiled the final version of the article and pushed it on the repo.
