### 2021-09-02 #54 @khinsen
Thanks @MathieuBo for the quick update. The paper is now published: https://zenodo.org/record/5379631/files/article.pdf
